moved both headers to Designer UI Header for consistency
