+++ Localize/lang/strings.json
@@ -948,7 +948,6 @@
   "Xkt2vD": "Select a function app function",
   "Xnn0uj": "Request",
   "XqamWZ": "Delete",
-  "Xs7Uvt": "By Microsoft",
   "XsktQ/": "Limit Logic Apps to not include workflow metadata headers in the response.",
   "XtuP5e": "Math functions",
   "Xv5CGN": "(UTC-05:00) Indiana (East)",
@@ -1957,7 +1956,6 @@
   "_Xkt2vD.comment": "Label for function app selection",
   "_Xnn0uj.comment": "Header text for request",
   "_XqamWZ.comment": "Label of Delete Token Button",
-  "_Xs7Uvt.comment": "Panel description for stating it was created by Microsoft",
   "_XsktQ/.comment": "description of workflow headers on response setting",
   "_XtuP5e.comment": "Label for math functions",
   "_Xv5CGN.comment": "Time zone value ",
