+++ libs/designer/src/lib/ui/panel/templatePanel/panel.less
@@ -3,12 +3,13 @@
     flex-direction: column;
     padding: 10px 0 20px 0;
     width: 100%;
-}
 
-.msla-template-panel-header {
-    font-size: 22px !important;
-    font-weight: 600 !important;
-    padding-bottom: 10px;
+    .msla-template-panel-header {
+        font-size: 22px;
+        line-height: 25px;
+        font-weight: 600;
+        padding-bottom: 10px;
+    }
 }
 
 .msla-template-quickview-tags {
