+++ libs/designer/src/lib/ui/panel/templatePanel/quickViewPanel/quickViewPanel.tsx
@@ -4,7 +4,7 @@ import { Text } from '@fluentui/react-components';
 import { Icon } from '@fluentui/react';
 import { useIntl } from 'react-intl';
 import { useState } from 'react';
-import { TemplatesPanelContent } from '@microsoft/designer-ui';
+import { TemplatesPanelContent, TemplatesPanelHeader } from '@microsoft/designer-ui';
 import { getQuickViewTabs } from '../../../../core/templates/utils/helper';
 import Markdown from 'react-markdown';
 
@@ -46,23 +46,24 @@ export const QuickViewPanelHeader = ({
   details,
 }: { title: string; description: string; details: Record<string, string> }) => {
   return (
-    <div className="msla-template-quickview-header">
-      <Text className="msla-template-panel-header">{title}</Text>
-      <div className="msla-template-quickview-tags">
-        {Object.keys(details).map((key: string, index: number, array: any[]) => {
-          return (
-            <div key={key}>
-              <Text className="msla-template-card-tag">
-                {key}: {details[key]}
-              </Text>
-              {index !== array.length - 1 ? (
-                <Icon style={{ padding: '3px 10px 3px 10px', color: '#dedede', fontSize: 10 }} iconName="LocationDot" />
-              ) : null}
-            </div>
-          );
-        })}
+    <TemplatesPanelHeader title={title}>
+      <div className="msla-template-quickview-header">
+        <div className="msla-template-quickview-tags">
+          {Object.keys(details).map((key: string, index: number, array: any[]) => {
+            return (
+              <div key={key}>
+                <Text className="msla-template-card-tag">
+                  {key}: {details[key]}
+                </Text>
+                {index !== array.length - 1 ? (
+                  <Icon style={{ padding: '3px 10px 3px 10px', color: '#dedede', fontSize: 10 }} iconName="LocationDot" />
+                ) : null}
+              </div>
+            );
+          })}
+        </div>
+        <Markdown linkTarget="_blank">{description}</Markdown>
       </div>
-      <Markdown linkTarget="_blank">{description}</Markdown>
-    </div>
+    </TemplatesPanelHeader>
   );
 };
